Remove the baseSeat class
